Rename morpheus.utils.logging to morpheus.utils.logger allowing other modules in morpheus.utils to import the std logger
