+++ morpheus/stages/general/delay_stage.py
@@ -20,7 +20,7 @@ import srf
 from morpheus.config import Config
 from morpheus.pipeline.single_port_stage import SinglePortStage
 from morpheus.pipeline.stream_pair import StreamPair
-from morpheus.utils.logging import deprecated_stage_warning
+from morpheus.utils.logger import deprecated_stage_warning
 
 logger = logging.getLogger(__name__)
